DSEGOG-67 refactoring plot settings tests
Still to get plot settings controller's snapshot right as we need to know the mock functions are being passed correctly
